ml_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
numpy==1.26.4
tensorflow==2.16.1
opencv-python-headless==4.9.0.80
pillow==10.3.0
python-multipart==0.0.9
httpx==0.27.0